% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poligono.R
\name{area}
\alias{area}
\title{Área}
\usage{
area(objeto)
}
\arguments{
\item{objeto}{Um objeto do tipo poligono.}
}
\value{
A área da figura geométrica.
}
\description{
Calcula a área de uma figura geométrica
}
\details{
Efetua o cálculo da área de uma figura geométrica baseado nas dimenções dos
segmentos que representam os lados da figura.
}
