% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operacoes.R
\name{desvio_padrao}
\alias{desvio_padrao}
\title{Desvio padrão}
\usage{
desvio_padrao(x)
}
\arguments{
\item{x}{Um vetor numérico para o qual se deseja calcular o desvio padrão}
}
\value{
O valor do desvio padrão dos elementos
}
\description{
Método que calcula o desvio padrão dos elementos de um vetor
}
