% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poligono.R
\name{perimetro}
\alias{perimetro}
\title{Perímetro}
\usage{
perimetro(objeto)
}
\arguments{
\item{objeto}{Um objeto do tipo poligono.}
}
\value{
O perímetro da figura geométrica.
}
\description{
Calcula o perímetro de uma figura geométrica
}
\details{
Efetua o cálculo do perímetro de uma figura geométrica baseado no número
de lados e das dimenções dos segmentos que representam os lados da figura.
}
