% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poligono.R
\name{poligono}
\alias{poligono}
\title{Polígono}
\usage{
poligono(lados)
}
\arguments{
\item{lados}{Define o número de lados da figura geométrica.}
}
\value{
Um novo objeto do tipo poligono
}
\description{
Representa um polígono
}
\details{
Define um objeto da geometria que é  uma figura plana formada por um número
finito de segmentos de reta conectados para formar seus lados.
}
