% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retangulo.R
\name{retangulo}
\alias{retangulo}
\title{Retangulo}
\usage{
retangulo(altura, largura)
}
\arguments{
\item{altura}{Define a altura da figura geométrica.}

\item{largura}{Define a largura da figura geométrica.}
}
\value{
Um novo objeto do tipo poligono
}
\description{
Representa um polígono com quatro lados
}
\details{
Define um objeto da geometria que é uma figura plana formada por um quatro
lados, sendo os lados paralelos de mesmo tamanho.
}
